+++ assets/js/app.js
@@ -28,15 +28,9 @@ import { LiveSocket } from 'phoenix_live_view';
 
 import topbar from '../vendor/topbar.cjs';
 import * as Hooks from './hooks';
-import JobEditor from './job-editor';
-import WorkflowEditor from './workflow-editor';
-import DataclipViewer from './dataclip-viewer';
 import LogViewer from './log-viewer';
 
 let hooks = {
-  JobEditor,
-  WorkflowEditor,
-  DataclipViewer,
   LogViewer,
   ...Hooks,
 };
@@ -123,6 +117,6 @@ window.liveSocket = liveSocket;
 
 // Testing helper to simulate a reconnect
 window.triggerReconnect = function triggerReconnect(timeout = 5000) {
-  liveSocket.disconnect(() => {});
+  liveSocket.disconnect(() => { });
   setTimeout(liveSocket.connect.bind(liveSocket), timeout);
 };